python_data_api/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
PyJWT==2.6.0
python-dotenv==0.21.0
Werkzeug==2.2.2
